+++ js/img-effects.js
@@ -1,6 +1,6 @@
 // Реализовал самостоятельно, лайв не смотрел, возможны ошибки. Хочу доработать основной обработчик, чтобы избавиться от switch-case. Хочу попробовать Array.find() по "evt.target.value".
 
-import './form.js';
+import './form-validate.js';
 import { EFFECTS } from './filters-for-noUi.js';
 
 // Глобальные переменные, нужны для присвоения стилей в зависимости от выбранного фильтра
