+++ js/main.js
@@ -1,12 +1,13 @@
 // import { photos } from './setup.js';
 
 // import { openBigPicture } from './full-size-pictures.js';
-import './form.js';
+import './form-validate.js';
 import './img-effects.js';
 import './scale-control.js';
 import './server-data.js';
 import { renderRemoteThumbnails, openCard } from './server-data.js';
-import { submitForm, closeForm } from './form.js';
+import { submitForm } from './form-upload.js';
+import { closeForm } from './callbacks.js';
 import './user-photo.js';
 
 renderRemoteThumbnails();
