+++ js/utils.js
@@ -1,4 +1,4 @@
-// Модуль вспомогательных функций
+
 
 function generateRandomInteger(firstInt, secondInt) {
   const min = Math.ceil(Math.min(firstInt, secondInt));
@@ -19,37 +19,17 @@ function getRandomIndex(arr) {
 }
 
 function debounce(callback, timeoutDelay = 500) {
-  // Используем замыкания, чтобы id таймаута у нас навсегда приклеился
-  // к возвращаемой функции с setTimeout, тогда мы его сможем перезаписывать
   let timeoutId;
-
   return (...rest) => {
-    // Перед каждым новым вызовом удаляем предыдущий таймаут,
-    // чтобы они не накапливались
     clearTimeout(timeoutId);
-
-    // Затем устанавливаем новый таймаут с вызовом колбэка на ту же задержку
     timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
-
-    // Таким образом цикл «поставить таймаут - удалить таймаут» будет выполняться,
-    // пока действие совершается чаще, чем переданная задержка timeoutDelay
   };
 }
 
 function throttle(callback, delayBetweenFrames) {
-  // Используем замыкания, чтобы время "последнего кадра" навсегда приклеилось
-  // к возвращаемой функции с условием, тогда мы его сможем перезаписывать
   let lastTime = 0;
-
   return (...rest) => {
-    // Получаем текущую дату в миллисекундах,
-    // чтобы можно было в дальнейшем
-    // вычислять разницу между кадрами
     const now = new Date();
-
-    // Если время между кадрами больше задержки,
-    // вызываем наш колбэк и перезаписываем lastTime
-    // временем "последнего кадра"
     if (now - lastTime >= delayBetweenFrames) {
       callback.apply(this, rest);
       lastTime = now;
@@ -57,6 +37,11 @@ function throttle(callback, delayBetweenFrames) {
   };
 }
 
-export { generateRandomInteger, uniqueID, getRandomIndex, debounce, throttle };
+const renderSuccessModal = () => {
+  document.body.append(successMessage);
+  successMessageModal.classList.add('hidden');
+};
+
+export { generateRandomInteger, uniqueID, getRandomIndex, debounce, throttle, renderSuccessModal };
